Move Random to RandomUtils
