+++ docs/examples/en/math/RandomUtils.html
@@ -24,7 +24,7 @@
 		<h2>Source</h2>
 
 		<p>
-		  [link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/math/Random.js examples/jsm/math/Random.js]
+		  [link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/math/RandomUtils.js examples/jsm/math/RandomUtils.js]
 		</p>
 	</body>
 </html>
