+++ examples/jsm/math/RandomUtils.js
